CHG: router 설정
