+++ jshong0907/auth/jwt.go
@@ -1,0 +1,9 @@
+package auth
+
+import "github.com/golang-jwt/jwt"
+
+type JwtClaims struct {
+	jwt.StandardClaims
+	Email    string `json:"email"`
+	NickName string `json:"NickName"`
+}
